Bug fix 1035 (partial)
I changed the call to remove a wme to tell it's parent wmobject to remove it, so that, if the removed wme is and identifier and therefore represented by a wmobject, the children attached to the wme can be removed as well.

SML needs to know when wmes are being removed out from underneath it. So, on agent creation, the gSKI agent is given a callback to use when removing elements in this case. Functions were added to KernelSML (and AgentSML) to allow this callback to update the records kept on the SML side.

Tests updated to reflect new changes. Tests for this moved to after stuff they depend on has been tested.

More details on bugzilla for bugs 1034, 1035