Merge of timer branch. Improved timers throughout code. New timers are available to all components in shared/misc.h and use the stlsoft library wrapped in a soar_timer interface, two soar_x_timer implementations (one for real time and the other for cpu time), and a soar_timer_accumulator for tracking multiple intervals. Stats and time commands updated to use the new system, and sml::OSSpecificTimer removed, code using it refactored to use the new code.
The timer code still needs to be commented, and a decision needs to be made whether or not to default to wall clock time as opposed to cpu time on all platforms because the resolution (at least on windows) is much, much higher for wall time. Right now, windows uses wall time for both, unix still uses cpu time when asked to.