first steps in module reorganization: 1. soar_module has basic infrastructure for calling module listeners at useful times 2. RL has been ported over to this organization, which removes a lot of its tentacles throughout the kernel 3. RL also has been OOPified and many variables converted from heap->stack 4. RL identifier variables have been moved to a private map, since they were only necessary on states
RL passes the Soar test suite, as well as runs all RL agents (water-jug, rl-unit, left-right). I have included in this commit an update to Joseph's last print commit to fix weird buffer overflow issues on Windows, but I'll wait for him/others to confirm/improve before porting to the trunk.

It's likely some of the module organization will change as we port epmem, smem, wma, etc., but this seems like a good start. Performance comparison to the trunk need to be a part of this process as well.