I, er, think I've found a bug in the selection space rules. Please double check this.
The change here is that the total-estimated-cost should be updated when estimated-cost != 0, and final-cost should be updated when estimated-cost == 0.
This is changed from the original code where the total-estimated-cost is updated if path-cost != 0, which would fail if start == goal, where path-cost == estimated-cost == total-estimated-cost == final-cost == 0.
The change makes sense when you think about A* - when the estimated cost is 0, that means you've reached the desired state, so the final cost is known.
Otherwise, the total cost is still an estimate, so the final cost is not stated and it's only a total *estimated* cost.

Side note: I think it's safe to put estimated-cost > 0 instead of estimated-cost != 0, since A* does weird things (read: doesn't work) on negative heuristic cost estimates.
I will however leave the change for future interested code readers (hi y'all!).