Turns out that, in order to simulate walls, a picture of a wall must be present.
Further more, if this picture does not actually exist, the simulated walls cannot be drawn in the viewer. This means that the file path of the picture of the wall is hard-coded into the Java code... the file path that was deleted in r13689.

This was also the case for a "skybox" and for floors; luckily, these are less essential to a simulation, and the idiocy of hard-coded .jpg file-paths was swiftly remedied.