Working patrol worst-preference.
The patrol code that I wrote in r/13700 doesn't work, as the operator retracts whenever the location changes. And yet it still works... What I *think* is going on is that a go-to operator is proposed (see random-walk/go-to-next-area.soar), and that somehow intelligently picks the next area to visit. Since my map is a loop, and since it avoids backtracking, it picks the right room.

I actually don't much care whether this explanation is right (well...), but it does work, and I'm moving on.