Epmem will now lookup constants from the four symbol lookup tables instead of the old temporal symbol hash table.  First working version.  Print and visualize episode code still hasn't been updated though yet.
How symbol type is handled is changed somewhat  from how smem does it and epmem previously did it;  that code retrieves the symbol types within the original large query. The theory behind this change is that it we'll gain more from removing a join from the big, more computationally intensive query than we lose from the overhead of a second query.  This leverages the fact that id's and attributes are always a known type, so we don't even need to join with the type table for those. (and reverse hashing for attributes is probably more common than for values with the way the code/tables are structured.)  We'll see if there's any truth to this when testing.