Fixed epmem storage bug
By the current logic in _epmem_remove_wme in rete.cpp, when a WME
is removed, the algorithm will recurse onto its children if the
value does not have any incoming WMEs from top state.

This is a problem, however, if the removed WMEs form a Y. Since
none of them have a top state TC, the algorithm will proceed depth
first, invalidating the id of the leg of the Y before the second
arm. When the arm gets removed, it's now-invalid epmem-id is returned
to the pool - causing problems when that id is reused.

The fix is that, instead of checking the TC of incoming edges, the
algorithm now only recurses if the ref_count is zero. This is a
stricter condition than the above, since if the ref_count is zero,
all incoming WMEs would (trivially) not be on the top state. Running
SoarTech's test cases on this fix suggests it does not cause
additional problems.