fixed working memory activation bug
The problem here is that the test for the o-support set of a WME is not sufficiently tight. In particular, if one of the tested WMEs has i-support, but *that i-support is removed at the same time as a rule that depends on it fires*, the result of that rule will have that i-supported WME in it's o-set. The previous test cannot catch this, since when the i-supported WME is removed, its preference is set to NIL. The fix is to also check the reference_count of the WME, to make sure that it has not already been removed from working memory. This is probably not fool proof; a different criteria will have to found if this is indeed the case.